packages/api-func/api-func/requirements.txt
fastapi==0.85.1
ibis-framework[postgres]==3.2.0
psycopg2==2.9.4
python-dotenv==0.21.0
sqlalchemy==1.4.42
uvicorn==0.19.0
requests==*